[INJIWEB-1564] make passcode template to handle onBack
Other changes
- rename PasscodeTemplate.tsx to PasscodePageTemplate.tsx
